WIP: refactor schedules page
