+++ src/lib/components/namespace-list.svelte
@@ -68,7 +68,7 @@
       {#each namespacesResult.filter( ({ namespace }) => namespace.includes(searchValue), ) as namespace}
         <li
           class="first:rounded-t-md first:border-t last:rounded-b-md border-b border-l border-r p-3 flex border-collapse gap-2 hover:bg-gray-50 cursor-pointer"
-          on:click={() => namespace?.onClick()}
+          on:click={() => namespace?.onClick(namespace.namespace)}
         >
           <div class="w-6 h-6 pl-3 active">
             {#if namespace.namespace === $page.params?.namespace}
@@ -76,7 +76,7 @@
             {/if}
           </div>
           <a
-            href={namespace.href}
+            href={namespace.href(namespace.namespace)}
             class="link"
             class:active={namespace.namespace === $page.params?.namespace}
             >{namespace.namespace}</a
